requirements.txt
absl-py==0.11.0
antialiased-cnns==0.3
autopep8==1.5.4
brewer2mpl==1.4.1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
dataclasses==0.8
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.10
imageio==2.9.0
importlib-metadata==3.1.0
install==1.3.4
joblib==0.17.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.4
oauthlib==3.1.0
Pillow==8.0.1
prettytable==2.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pynvml==8.0.4
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.54.0
typing-extensions==3.7.4.3
uncertainty-metrics==0.0.81
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.0
